Split GraphQL logic in multiple files
